+++ build-pkg/build_n_check.R
@@ -42,7 +42,7 @@ pkg_loc <- build(path="build-pkg")
 
 # Test the package
 ## Sys.setenv(LANG="en_US.iso88591")
-Sys.setlocale("LC_ALL", "en_US.iso88591")
+## Sys.setlocale("LC_ALL", "en_US.iso88591")
 Sys.setenv('_R_CHECK_SYSTEM_CLOCK_' = 0)
 check_built(path=pkg_loc)