+++ data-raw/solving_issues.R
@@ -3,10 +3,4 @@
 # Author: Miguel Alvarez
 ################################################################################
 
-library(vegtable)
-
-load_last("../../sam_wetlands/syntax_chile_db/data/cl_syntax")
-
-summary(cl_syntax@species)
-
-object <- cl_syntax@species
+# Recycle