+++ NEWS.md
@@ -1,6 +1,10 @@
 taxlist 0.2.1
 =============
 
+### New Features
+
+* New function `indented_list()` to print taxonomic ranks in indented lists.
+
 ### Improvements
 
 * New argument `repaste` in function `dissect_name()` for re-pasting